Prepare to jump into init process
