+++ cfg/grub.cfg
@@ -3,5 +3,6 @@ set default=0
 
 menuentry "FlowerOS" {
     multiboot2 /boot/kernel.elf
+    module2 --nounzip  /boot/init.elf
     boot
 }
