+++ init/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "init"
+version = "0.1.0"
+authors = ["Restioson <restiosondev@gmail.com>"]
+edition = "2018"
+
+[lib]
+crate-type = ["staticlib"]
+
+[profile.release]
+panic = "abort"
+
+[profile.dev]
+panic = "abort"
+opt-level = 3
+
+[dependencies]
