+++ init/src/asm/start.asm
@@ -1,0 +1,8 @@
+extern main
+global start
+
+section .text.start
+bits 64
+
+start:
+    call main
