+++ init/src/lib.rs
@@ -1,0 +1,10 @@
+#![no_std]
+
+#![feature(lang_items)]
+
+mod lang;
+
+#[no_mangle]
+pub extern "C" fn main() {
+    loop {}
+}
