+++ init/x86_64-unknown-flower-none.json
@@ -1,0 +1,15 @@
+{
+  "arch": "x86_64",
+  "cpu": "x86-64",
+  "data-layout": "e-m:e-i64:64-f80:128-n8:16:32:64-S128",
+  "llvm-target": "x86_64-unknown-none-none",
+  "linker-flavor": "gcc",
+  "no-compiler-rt": true,
+  "os": "flower",
+  "target-endian": "little",
+  "target-pointer-width": "64",
+  "target-c-int-width": "32",
+  "features": "-mmx,-sse,-sse2,-sse3,-ssse3,-sse4.1,-sse4.2,-sse4a,-3dnow,-3dnowa,-avx,-avx2,+soft-float",
+  "disable-redzone": true,
+  "eliminate-frame-pointer": false
+}
